Moved Validate and RegExPatterns from Common to Web
