+++ src/Dfc.CourseDirectory.Web/Controllers/BulkUploadApprenticeshipsController.cs
@@ -17,6 +17,7 @@ using Dfc.CourseDirectory.Services.Interfaces.BulkUploadService;
 using Dfc.CourseDirectory.Services.Interfaces.CourseService;
 using Dfc.CourseDirectory.Services.Interfaces.ProviderService;
 using Dfc.CourseDirectory.Web.Helpers;
+using Dfc.CourseDirectory.Web.Validation;
 using Dfc.CourseDirectory.Web.ViewModels.BulkUpload;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -132,7 +133,7 @@ namespace Dfc.CourseDirectory.Web.Controllers
                 bulkUploadFile.CopyTo(ms);
 
                 ms.Position = 0;
-                if (Validate.isBinaryStream(ms))
+                if (Validate.IsBinaryStream(ms))
                 {
                     return View(new BulkUploadViewModel {errors = new[] {"Invalid file content."}});
                 }
