+++ src/Dfc.CourseDirectory.Web/ViewComponents/CourseName/Default.cshtml
@@ -1,5 +1,5 @@
 ﻿@using System.Net
-@using Dfc.CourseDirectory.Common
+@using Dfc.CourseDirectory.Web.Validation
 @model Dfc.CourseDirectory.Web.ViewComponents.CourseName.CourseNameModel
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
