+++ src/Dfc.CourseDirectory.Web/ViewComponents/CourseProviderReference/Default.cshtml
@@ -1,5 +1,5 @@
 ﻿@using System.Net
-@using Dfc.CourseDirectory.Common
+@using Dfc.CourseDirectory.Web.Validation
 @model Dfc.CourseDirectory.Web.ViewComponents.CourseProviderReference.CourseProviderReferenceModel
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
@@ -32,7 +32,7 @@
 
 
 
-    <script>
+<script>
     (function ($) {
         $(function () {
             var $elementCourseProviderReference = $("#@nameof(Model.CourseProviderReference)");
@@ -43,4 +43,4 @@
             });
         });
     })(jQuery);
-    </script>
+</script>
