+++ src/Dfc.CourseDirectory.Web/ViewComponents/Courses/WhatWillLearn/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using Dfc.CourseDirectory.Common
+﻿@using Dfc.CourseDirectory.Web.Validation
 @using Dfc.CourseDirectory.Common.Settings
 @using Microsoft.Extensions.Options
 @model Dfc.CourseDirectory.Web.ViewComponents.Courses.WhatWillLearn.WhatWillLearnModel
@@ -9,7 +9,7 @@
 
 
 @{
- @*   var regex = @"^[a-zA-Z0-9 /\n/\r/\\u/\¬\!\£\$\%\^\&\*\\é\\è\\ﬁ\(\)_\+\-\=\{\}\[\]\;\:\@\'\#\~\,\<\>\.\?\/\|\`\•\·\●\\’\‘\“\”\—\-\–\‐\‐\…\:/\°\®\\â\\ç\\ñ\\ü\\ø\♦\™\\t/\s\¼\¾\½\" + "\"" + "\\\\]+$";*@
+    @*   var regex = @"^[a-zA-Z0-9 /\n/\r/\\u/\¬\!\£\$\%\^\&\*\\é\\è\\ﬁ\(\)_\+\-\=\{\}\[\]\;\:\@\'\#\~\,\<\>\.\?\/\|\`\•\·\●\\’\‘\“\”\—\-\–\‐\‐\…\:/\°\®\\â\\ç\\ñ\\ü\\ø\♦\™\\t/\s\¼\¾\½\" + "\"" + "\\\\]+$";*@
     var regex = RegexPattern.AllowEverything;
     var whatWillLearnTextFieldMaxChars = whatWilllLearnComponentSettings.Value.TextFieldMaxChars;
 }
@@ -28,7 +28,7 @@
             </div>
         }
         <textarea class="govuk-textarea js-character-count " asp-for="@Model.WhatWillLearn" rows="7"
-                    aria-describedby="@Model.AriaDescribedBy" aria-label="WhatYouWillLearn"
+                  aria-describedby="@Model.AriaDescribedBy" aria-label="WhatYouWillLearn"
                   val-max-length=@whatWillLearnTextFieldMaxChars val-max-length-message="What you will learn must be @whatWillLearnTextFieldMaxChars characters or less"
                   val-regex="@regex"
                   val-regex-message="What you’ll learn contains invalid characters"></textarea>
