+++ src/Dfc.CourseDirectory.Web/ViewComponents/Courses/WhereNext/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using Dfc.CourseDirectory.Common
+﻿@using Dfc.CourseDirectory.Web.Validation
 @using Dfc.CourseDirectory.Common.Settings
 @using Microsoft.Extensions.Options
 @model Dfc.CourseDirectory.Web.ViewComponents.Courses.WhereNext.WhereNextModel
@@ -10,7 +10,7 @@
 
 
 @{
-@*    var regex = @"^[a-zA-Z0-9 /\n/\r/\\u/\¬\!\£\$\%\^\&\*\\é\\è\\ﬁ\(\)_\+\-\=\{\}\[\]\;\:\@\'\#\~\,\<\>\.\?\/\|\`\•\·\●\\’\‘\“\”\—\-\–\‐\‐\…\:/\°\®\\â\\ç\\ñ\\ü\\ø\♦\™\\t/\s\¼\¾\½\" + "\"" + "\\\\]+$";*@
+    @*    var regex = @"^[a-zA-Z0-9 /\n/\r/\\u/\¬\!\£\$\%\^\&\*\\é\\è\\ﬁ\(\)_\+\-\=\{\}\[\]\;\:\@\'\#\~\,\<\>\.\?\/\|\`\•\·\●\\’\‘\“\”\—\-\–\‐\‐\…\:/\°\®\\â\\ç\\ñ\\ü\\ø\♦\™\\t/\s\¼\¾\½\" + "\"" + "\\\\]+$";*@
     var regex = RegexPattern.AllowEverything;
     var whereNextTextFieldMaxChars = whereNextComponentSettings.Value.TextFieldMaxChars;
 }
@@ -30,9 +30,9 @@
         }
         <textarea class="govuk-textarea js-character-count " asp-for="@Model.WhereNext" rows="7"
                   aria-describedby="@Model.AriaDescribedBy" aria-label="WhereNext"
-                          val-max-length=@whereNextTextFieldMaxChars val-max-length-message="'Where next' must be @whereNextTextFieldMaxChars characters or less"
-                          val-regex="@regex"
-                          val-regex-message="Where next contains invalid characters"></textarea>
+                  val-max-length=@whereNextTextFieldMaxChars val-max-length-message="'Where next' must be @whereNextTextFieldMaxChars characters or less"
+                  val-regex="@regex"
+                  val-regex-message="Where next contains invalid characters"></textarea>
 
     </div>
 </div>
